+++ src/test/java/com/iota/iri/model/HashTest.java
@@ -31,7 +31,7 @@ public class HashTest {
 
     @Test
     public void calculate2() throws Exception {
-        int[] trits = TransactionViewModelTest.getRandomTransactionTrits();
+        byte[] trits = TransactionViewModelTest.getRandomTransactionTrits();
         byte[] bytes = Converter.allocateBytesForTrits(trits.length);
         Converter.bytes(trits, bytes);
         Hash hash = Hash.calculate(bytes, TransactionViewModel.TRINARY_SIZE, SpongeFactory.create(SpongeFactory.Mode.CURLP81));
@@ -49,12 +49,12 @@ public class HashTest {
     @Test
     public void trits() throws Exception {
         Hash hash = Hash.calculate(SpongeFactory.Mode.CURLP81, TransactionViewModelTest.getRandomTransactionTrits());
-        Assert.assertFalse(Arrays.equals(new int[Hash.SIZE_IN_TRITS], hash.trits()));
+        Assert.assertFalse(Arrays.equals(new byte[Hash.SIZE_IN_TRITS], hash.trits()));
     }
 
     @Test
     public void equals() throws Exception {
-        int[] trits = TransactionViewModelTest.getRandomTransactionTrits();
+        byte[] trits = TransactionViewModelTest.getRandomTransactionTrits();
         Hash hash = Hash.calculate(SpongeFactory.Mode.CURLP81, trits);
         Hash hash1 = Hash.calculate(SpongeFactory.Mode.CURLP81, trits);
         Assert.assertTrue(hash.equals(hash1));
@@ -64,7 +64,7 @@ public class HashTest {
 
     @Test
     public void hashCodeTest() throws Exception {
-        int[] trits = TransactionViewModelTest.getRandomTransactionTrits();
+        byte[] trits = TransactionViewModelTest.getRandomTransactionTrits();
         Hash hash = Hash.calculate(SpongeFactory.Mode.CURLP81, trits);
         Assert.assertNotEquals(hash.hashCode(), 0);
         Assert.assertEquals(Hash.NULL_HASH.hashCode(), -240540129);
@@ -72,7 +72,7 @@ public class HashTest {
 
     @Test
     public void toStringTest() throws Exception {
-        int[] trits = TransactionViewModelTest.getRandomTransactionTrits();
+        byte[] trits = TransactionViewModelTest.getRandomTransactionTrits();
         Hash hash = Hash.calculate(SpongeFactory.Mode.CURLP81, trits);
         Assert.assertEquals(Hash.NULL_HASH.toString(), "999999999999999999999999999999999999999999999999999999999999999999999999999999999");
         Assert.assertNotEquals(hash.toString(), "999999999999999999999999999999999999999999999999999999999999999999999999999999999");
@@ -82,7 +82,7 @@ public class HashTest {
 
     @Test
     public void bytes() throws Exception {
-        int[] trits = TransactionViewModelTest.getRandomTransactionTrits();
+        byte[] trits = TransactionViewModelTest.getRandomTransactionTrits();
         Hash hash = Hash.calculate(SpongeFactory.Mode.CURLP81, trits);
         Assert.assertTrue(Arrays.equals(new byte[Hash.SIZE_IN_BYTES], Hash.NULL_HASH.bytes()));
         Assert.assertFalse(Arrays.equals(new byte[Hash.SIZE_IN_BYTES], hash.bytes()));
@@ -91,7 +91,7 @@ public class HashTest {
 
     @Test
     public void compareTo() throws Exception {
-        int[] trits = TransactionViewModelTest.getRandomTransactionTrits();
+        byte[] trits = TransactionViewModelTest.getRandomTransactionTrits();
         Hash hash = Hash.calculate(SpongeFactory.Mode.CURLP81, trits);
         Assert.assertEquals(hash.compareTo(Hash.NULL_HASH), -Hash.NULL_HASH.compareTo(hash));
     }
